Clarify definition.
Putting `Point;` on the same line as the closing brace makes it easier to
see at a glance that it is part of the `typedef struct` definition. It
is the part which actually tells the compiler that you want to be able
to refer to the `struct Point` as just `Point`.