Minimize variable scope and give meaningful names
Variables used in only one place should generally not be defined in the
global scope. A good general principle is to always limit the scope of
any data as much as possible because it simplifies understanding what
effects changing the data could have by reducing the number of places
one needs to look to find where it is used.