Simplify reading of start button.
A few things going on here.

1. It looks like `startButton` was intended to hold the pin number of
the index pin, but then it was used to hold the value. In the end it
works but it would cause confusion for someone reading the code.
2. Constants, like the number of the pin used for the start button,
are generally named with all caps. The convention tells a reader that
the value should not change at a glance.
3. Constants can be `#define`ed but it is better to mark them with
`const` because it lets the compiler understand what you are trying to
do better, and that lets it give you better error messages when things
go wrong.
4. In C/C++ you can give a number to anything which expects a boolean
and the value will be understood as "true" if it is nonzero. Sometimes
it is still clearer to explicitly compare it to a number but in this
case it's probably clearer to directly feed the `digitalRead` to the
`while` loop.