Simplify setting pen down and increasing the speed for drawing
Problem: move fast to the place where the drawing is going to start,
then set the pen down and start moving faster.
Previous solution: increment a counter within the inner drawing loop and
set the speed high if it is nonzero. Continuously set the pen down after
moving.
New solution: create functions to raise and lower the pen which also
increase or decrease the motor speed for moving or drawing. Keep track
of whether drawing is happening in a `mode` variable which can be
checked to know whether to lower the pen.